comment auth subtleties, bound auth-retry to 1, split kb-session provider into modules
